Initial server code
This is the initial code for Clicker.  Clicker is a remote control
XML-RPC server that allows one to control home theater devices
that have a rs232 or HTTP interface.  It should also be possible to
use IR by integrating with LIRC.

Clicker has a concept of activities that can be defined.  Each
activity can power on the appropriate devices and set the proper
inputs.  This allows a one-button approach to common tasks, such
as watching a DVD, or listening to a CD.

The code is currently able to do the following:

- load controllable device config files
- load activity config files
- list devices via XML-RPC
- list device buttons via XML-RPC
- press buttons via XML-RPC
- list activities
- start activities via XML-RPC
- send commands to HTTP controlled devices

The rs232 commands are not currently sent to the devices yet, as
testing needs to be done with my specific devices.  The HTTP mode
has been smoke tested with a DTV HR21 receiver.  Retrieving and
parsing status command responses still needs to be added as well.

There is a test client script which shows the basics for testing
the XML-RPC interface from Python.  A CLI is planned for the future
along with a GUI client.