Add ability to run as a daemon
This makes the server run as a daemon.  There are details in the
README.md file that define how a system should be configured.
It is intended that the server is started as root and drops privs
to an unpriviledged user.  The user/group to run as are specified
in the config file.  A pidfile is also specified in the config file
which is used to ensure that only one server process is running.

The server and client source files were renamed to match the name
of the project.