Allow http timeout to be set per-device
Previously, the timeout for http connections to managed devices
was just whatever the global default was.  This patch adds the
ability to put a timeout setting in the http section of a device
config file that uses the http communication mode.  This timeout
will be used for all http requests that we send to that device.

I also fixed an error where the urllib2 module name was not
specified when we tried to catch a a URLError.