Pass config file path as an a command-line argument
This removes the hardcoded config file path and instead requires
that it be passed as a positional argument on the command-line
when starting the server.

I also added a --debug flag that enables redirecting stderr and
stdout to a log file in /tmp after daemonizing.  If this flag is
not set, these are both redirected to /dev/null.  All logging is
output to the configured logfile, so this is only really helpful
for uncaught exceptions.