Add RPM packaging support
This adds the ability to create a RPM package.  In order to
accomplish this, I needed to add a standard Python build script
and a spec file that uses it to build.  This also was a good
time to restructure the source tree.

As part of providing a working RPM, I added all of the proper
installation logic to the spec file.  The clicker user and group
are created during RPM install, and all of the proper system dirs
are created with the proper permissions.  All that is needed after
installing is to create the activity and device config files and
start the daemon.

A systemd service file was added to allow stopping and starting
of the clickerd daemon.

Lastly, proper licensing of the code was done, specifically using
GPLv3+.  We install the license in the COPYING file in the clicker
documentation area.