Add a startup pipe for returning daemon startup status
This adds a startup pipe that is used to allow the daemon process
to report startup status to the parent process.  When starting, the
parent remains running after forking the child daemon process.  The
parent process opens up a pipe for reading.  The daemon process
opens up the write side of the pipe and outputs any startup errors
or a success message when it is ready to process client requests.
The parent process can then return a useful exit code that truly
indicates if the daemon started successfully.

This requires some changes to python-daemon, which I have made and
submitted upstream.  The changes allow one to pass a parent context
in when creating the DaemonContext.  This parent context is used by
the parent process after forking instead of having the parent exit
immediately as the current python-daemon does.  This parent context
is what we use to open the read side of the pipe after forking.