Add serial device button and status command support
This adds support for status commands for rs232 devices.  One can
define status commands and responses to be returned to client in
the device config files.  Filtering and mapping of the actual
response from the device into meaningful values can be performed.

In addition, button presses for serial devices now works as well.
One can specify a serial read timeout in the device config files,
otherwise the clicker daemon defaults to a 1 second timeout.