Add initial Android client
This adds the code for a Clicker Android app.  The app is functional
at this point, though there is still some polishing to be done.  It
currently displays controlled devices and has button lists for each
device that send the proper commands to the server.

The action bar has icons for setting the connection settings for
the server, refreshing the device lists, starting activities, and
a master power of button used to quit activities.