Clean up Eclipse project for Android client
The Clicker Android client portion of the tree had some files
checked in related to the Eclipse project that were build files.
These do not need to be in the source tree.

The Android v4 support library was also checked in, but one should
really use the library included with their local Android SDK.  The
classpath is set such that one needs to set the ANDROID_SDK variable
to point to the Android SDK location to locate the support library.