Set power switch initial state from device status
This allows the initial state of the power switch for a device in
the client application to be set based off of the power status
that is reported by a device.

In addition to changes required to the client application code, the
server code needed some changes to be able to read varying status
response lengths for a specific status command.  To do this, we
need to use io.TextIOWrapper to allow us to use readline() that
supports end-of-line characters other than \n.  If readline() is
to be used to get the status from a device, the device config file
must specify the size of the response as 0.

These changes work with the three rs232 controlled devices whose
config files are included in the source tree.  This has not been
made to work with HTTP controlled devices yet.