Add spinner view for input selection to client application
This adds a spinner view for selecting the input for a device
in the client application.  If a device has input_ buttons, we
build a list of those inputs and populate the values for a spinner
that is displayed instead of listing the input buttons in the main
button list.

If the device also has an input status command, the spinner
selection is updated to match the currently selected input when
creating the spinner.  This selection is updated when powering on
a device and selecting a device.